simulator/requirements.txt
# 无人机模拟器依赖
paho-mqtt>=1.6.0
requests>=2.25.0 